+++ install.sh
@@ -432,7 +432,7 @@ pushd .
 # #################################################
 # configure & build
 # #################################################
-cmake_common_options=""
+cmake_common_options="--toolchain=toolchain-linux.cmake"
 cmake_client_options=""
 
 if [[ "${build_address_sanitizer}" == true ]]; then
@@ -487,20 +487,18 @@ if [[ "${build_clients}" == true ]]; then
     cmake_client_options="${cmake_client_options} -DBUILD_CLIENTS_SAMPLES=ON -DBUILD_CLIENTS_TESTS=ON -DBUILD_CLIENTS_BENCH=ON"
 fi
 
-compiler="hipcc"
-
 # Build library with AMD toolchain because of existense of device kernels
 if [[ "${build_clients}" == false ]]; then
     cmake_client_options=" "
 fi
 if [[ "${build_relocatable}" == true ]]; then
-    CXX=${compiler} CC=${compiler} ${cmake_executable} ${cmake_common_options} ${cmake_client_options} -DCPACK_SET_DESTDIR=OFF -DCMAKE_INSTALL_PREFIX="${install_prefix}" -DCPACK_PACKAGING_INSTALL_PREFIX="${rocm_path}" \
-       -DCMAKE_PREFIX_PATH="${rocm_path} ${rocm_path}/hipcc ${rocm_path}/hip" \
+    ${cmake_executable} ${cmake_common_options} ${cmake_client_options} -DCPACK_SET_DESTDIR=OFF -DCMAKE_INSTALL_PREFIX="${install_prefix}" -DCPACK_PACKAGING_INSTALL_PREFIX="${rocm_path}" \
+       -DCMAKE_PREFIX_PATH="${rocm_path} ${rocm_path}/amdclang++ ${rocm_path}/hip" \
        -DCMAKE_SHARED_LINKER_FLAGS="${rocm_rpath}" \
        -DROCM_DISABLE_LDCONFIG=ON \
        ../..
 else
-    CXX=${compiler} CC=${compiler} ${cmake_executable} ${cmake_common_options} ${cmake_client_options} -DCPACK_SET_DESTDIR=OFF -DCMAKE_INSTALL_PREFIX=${install_prefix} -DCPACK_PACKAGING_INSTALL_PREFIX=/opt/rocm ../..
+    ${cmake_executable} ${cmake_common_options} ${cmake_client_options} -DCPACK_SET_DESTDIR=OFF -DCMAKE_INSTALL_PREFIX=${install_prefix} -DCPACK_PACKAGING_INSTALL_PREFIX=/opt/rocm ../..
 fi
 check_exit_code
 make -j$(nproc)