+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Documentation for rocFFT is available at
 
 ### Changes
 
+* Compile with amdclang++ instead of hipcc.
 * Add --smoketest option to rocfft-test.
 * Support gfx1200 and gfx1201 architectures.
 * Replace Boost Program Options with CLI11 as the command line parser for clients and samples.